requirements.txt
Django==6.0.3
djangorestframework==3.17.0
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
openai==2.29.0
python-dotenv==1.2.2
